已存在的ypk，yrp文件不做处理只给提示

+++ mobile/src/main/java/cn/garymb/ygomobile/GameUriManager.java
@@ -134,10 +134,14 @@ public class GameUriManager {
                 activity.startActivity(startdeck);
             } else if (file.getName().toLowerCase(Locale.US).endsWith(".ypk")) {
                 File ypk = new File(AppsSettings.get().getExpansionsPath() + "/" + file.getName().toLowerCase(Locale.US));
-                try {
-                    FileUtils.copyFile(file, ypk);
-                } catch (Throwable e) {
-                    Toast.makeText(activity, activity.getString(R.string.install_failed_bcos) + e, Toast.LENGTH_LONG).show();
+                if (ypk.exists()) {
+                    Toast.makeText(activity, activity.getString(R.string.file_exist), Toast.LENGTH_LONG).show();
+                } else {
+                    try {
+                        FileUtils.copyFile(file, ypk);
+                    } catch (Throwable e) {
+                        Toast.makeText(activity, activity.getString(R.string.install_failed_bcos) + e, Toast.LENGTH_LONG).show();
+                    }
                 }
                 if (!AppsSettings.get().isReadExpansions()) {
                     activity.startActivity(startSeting);
@@ -148,8 +152,9 @@ public class GameUriManager {
                 }
             } else if (file.getName().toLowerCase(Locale.US).endsWith(".yrp")) {
                 File yrp = new File(AppsSettings.get().getResourcePath() + "/" + CORE_REPLAY_PATH + "/" + file.getName().toLowerCase(Locale.US));
-                if (yrp.getAbsolutePath().equals(file)) {
+                if (yrp.exists()) {
                     YGOStarter.startGame(getActivity(), null);
+                    Toast.makeText(activity, activity.getString(R.string.file_exist), Toast.LENGTH_LONG).show();
                 } else {
                     try {
                         FileUtils.copyFile(file, yrp);
@@ -190,15 +195,19 @@ public class GameUriManager {
                 try {
                     File ypk = new File(AppsSettings.get().getExpansionsPath() + "/" + urifile.getName().toLowerCase(Locale.US));
                     ParcelFileDescriptor pfd = getActivity().getContentResolver().openFileDescriptor(uri, "r");
-                    if (pfd == null) {
-                        return;
+                    if (ypk.exists()) {
+                        Toast.makeText(activity, activity.getString(R.string.file_exist), Toast.LENGTH_SHORT).show();
                     } else {
-                        try {
-                            FileUtils.copyFile(new FileInputStream(pfd.getFileDescriptor()), ypk);
-                        } catch (Throwable e) {
-                            Toast.makeText(activity, activity.getString(R.string.install_failed_bcos) + e, Toast.LENGTH_LONG).show();
-                        } finally {
-                            pfd.close();
+                        if (pfd == null) {
+                            return;
+                        } else {
+                            try {
+                                FileUtils.copyFile(new FileInputStream(pfd.getFileDescriptor()), ypk);
+                            } catch (Throwable e) {
+                                Toast.makeText(activity, activity.getString(R.string.install_failed_bcos) + e, Toast.LENGTH_LONG).show();
+                            } finally {
+                                pfd.close();
+                            }
                         }
                     }
                 } catch (Throwable e) {
@@ -217,6 +226,7 @@ public class GameUriManager {
                     ParcelFileDescriptor pfd = getActivity().getContentResolver().openFileDescriptor(uri, "r");
                     if (yrp.exists()) {
                         YGOStarter.startGame(getActivity(), null);
+                        Toast.makeText(activity, activity.getString(R.string.file_exist), Toast.LENGTH_SHORT).show();
                     } else {
                         if (pfd == null) {
                             return;
